test: gate the launch-contract sentinels behind a non-default feature
The three falsification sentinels pin environment contracts — zebrad's
regtest behavior and zcash_local_net's launch contract — not this
repository's code, so no zls change can fail them and every default-
suite execution between environment changes re-verifies an unmoved
world. They move from tests/observability.rs into tests/sentinels.rs
behind the new non-default `sentinels` feature: run them when the
environment moves (infrastructure pin bumps, container-image bumps,
activation-height changes). The evidence test stays in the default
suite, where it exercises this repository's own scenario, replay, and
isolation machinery under real concurrency.

The manifest also carries the porting lane's concurrent `unit_test_twins`
declaration (pre-migration live originals of tests whose offline twins
now hold the default coverage) — included here because the new
non-default `extra-credit-tests` aggregate references it: one switch
for everything beyond the default suite.
