+++ libtonode-tests/Cargo.toml
@@ -11,6 +11,17 @@ default = ["chain_generic_tests", "proptests"]
 proptests = []
 chain_generic_tests = []
 ci = ["zingolib/ci"]
+# Pre-migration originals of tests that gained offline unit twins
+# (docs/testing/live-offline-twins.md). Off by default: the twins carry
+# the default coverage; enable to run the live originals side-by-side.
+unit_test_twins = []
+# Sentinel tests pin environment contracts (zebrad behavior and the
+# zcash_local_net launch contract), not this repository's code. Off by
+# default: run them when the environment moves — infra pin bumps,
+# container-image bumps, activation-height changes.
+sentinels = []
+# Everything beyond the default suite, in one switch.
+extra-credit-tests = ["sentinels", "unit_test_twins"]
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 [dependencies]
